requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
reportlab
pdf2image
pytesseract
Pillow
